Update RightTriangle.cpp
upd. Поменял int на double :).  I changed the int variable to double

+++ RightTriangle.cpp
@@ -3,23 +3,23 @@
 
 using namespace std;
 
-bool isTriangle(int a, int b, int c)
+bool isTriangle(double a, double b, double c)
 {
 	return (a + b > c and a + c > b and b + c > a);
 }
 
-bool isRightTriangle(int a, int b, int c)
+bool isRightTriangle(double a, double b, double c)
 {
-	int a2 = pow(a, 2);
-	int b2 = pow(b, 2);
-	int c2 = pow(c, 2);
+	double a2 = pow(a, 2);
+	double b2 = pow(b, 2);
+	double c2 = pow(c, 2);
 	return (a2 + b2 == c2 or a2 + c2 == b2 or b2 + c2 == a2);
 }
 
 int main()
 {
 	cout << "Enter the numbers for the triangle: " << endl;
-	int a, b, c;
+	double a, b, c;
 	cin >> a >> b >> c;
 
 	if (isRightTriangle(a, b, c))
